Fixed local variables not restored after <$content>
move_dlnode appended the node at the end of the destination varlist.

But find_varname scanned the list from the beginnen and stopped on first
match of variables name, ignoring mci. So if there was another (global)
instance in the varlist it always shadowed the local one after restoring
the varlist in handle_hsc_content.