Fixed tags outside any branch and added missing tags
Added tag to mark latest release 1.0b, but decided against adding tags
for versions 0.930 to 0.934 and 0.999 as it wasn't clear to which commit
to apply them.

Tags to mark releases 0.6.3 to 0.917 weren't reachable from any branch,
especially not from the master branch. So we've moved any tag to its
preceding "Updated to version" commit which is part of the history of
the master branch.

While this IS history rewriting (which is usually bad) and let us loose
commits like 11c218b (see reference below) this looked like the least
drastic method to fix the problem.

The commits lost are empty commits with the message "Added tag for
version $v" immediately following a commit with the message "Updated
to version $v.". And we had to tag the releases properly.

For reference, a 'git log --graph --oneline --all' showed us:

* cca01d3 Updated to version 0.8.1.
| * 2aac22d (tag: 0.7.5) Added tag for version 0.7.5.
|/
* 1210f07 Updated to version 0.7.5.
| * 11c218b (tag: 0.6.3) Added tag for version 0.6.3.
|/
* 06f1afd Added initial import of version 0.6.3.
* 7035d8e Added Subversion standard project directory structure.

Further a 'git show 0.6.3' showed us:

tag 0.6.3
Tagger: Matthias Bethke <matthias@towiski.de>
Date:   Wed Aug 30 19:46:11 1995 +0000

Added tag for version 0.6.3.

commit 11c218b8df7b7899868332f3337accad3ba8dabe (tag: 0.6.3)
Author: Thomas Aglassinger <roskakori@users.sourceforge.net>
Date:   Wed Aug 30 19:46:11 1995 +0000

    Added tag for version 0.6.3.